pyproject.toml
[tool.poetry]
name = "backend-pomodoro-task"
version = "0.1.0"
description = ""
authors = ["Alisson Oliveira <alisson@coopervap.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.1"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.1.0"
django-apscheduler = "^0.7.0"
apscheduler = "^3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
